pyproject.toml
[tool.poetry]
name = "taskmanager"
version = "0.1.0"
description = ""
authors = ["Artyom <petrakovartyom@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["all"], version = "^0.115.6"}
uvicorn = "^0.34.0"
psycopg2-binary = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
